Move PackageDefinition and PackageVersion out of module
