+++ cosmos-model/src/main/scala/com/mesosphere/universe/PackageVersion.scala
@@ -1,0 +1,34 @@
+package com.mesosphere.universe
+
+import com.twitter.util.Return
+import com.twitter.util.Throw
+import com.twitter.util.Try
+
+package v3.model {
+
+  sealed abstract class PackagingVersion private[model](val show: String)
+
+  object PackagingVersion {
+    val allVersions = Seq(V2PackagingVersion, V3PackagingVersion)
+
+    private[this] val allVersionsString = allVersions.map(_.show).mkString(", ")
+
+    private[this] val allVersionsIndex: Map[String, PackagingVersion] = {
+      allVersions.map(v => v.show -> v).toMap
+    }
+
+    def apply(s: String): Try[PackagingVersion] = {
+      allVersionsIndex.get(s) match {
+        case Some(v) => Return(v)
+        case _ => Throw(new IllegalArgumentException(
+          s"Expected one of [$allVersionsString] for packaging version, but found [$s]"
+        ))
+      }
+    }
+  }
+
+  case object V2PackagingVersion extends PackagingVersion("2.0")
+
+  case object V3PackagingVersion extends PackagingVersion("3.0")
+
+}
